Refactoring: create setting actions creator :dizzy:
