+++ mobile/src/actions/mediaBackup.js
@@ -6,8 +6,12 @@ export const MEDIA_UPLOAD_START = 'MEDIA_UPLOAD_START'
 export const MEDIA_UPLOAD_END = 'MEDIA_UPLOAD_END'
 export const IMAGE_UPLOAD_SUCCESS = 'IMAGE_UPLOAD_SUCCESS'
 
+export const startMediaUpload = () => ({ type: MEDIA_UPLOAD_START })
+export const endMediaUpload = () => ({ type: MEDIA_UPLOAD_END })
+export const successImageUpload = (image) => ({ type: IMAGE_UPLOAD_SUCCESS, id: image.id })
+
 export const mediaBackup = () => async (dispatch, getState) => {
-  dispatch({ type: MEDIA_UPLOAD_START })
+  dispatch(startMediaUpload())
   let photos = await getPhotos()
   const alreadyUploaded = getState().mobile.mediaBackup.uploaded
   for (let photo of photos) {
@@ -18,14 +22,14 @@ export const mediaBackup = () => async (dispatch, getState) => {
         name: photo.filename
       }
       await cozy.files.create(blob, options).then(() => {
-        dispatch({ type: IMAGE_UPLOAD_SUCCESS, id: photo.id })
+        dispatch(successImageUpload(photo))
       }).catch(err => {
         if (err.status === HTTP_CODE_CONFLICT) {
-          dispatch({ type: IMAGE_UPLOAD_SUCCESS, id: photo.id })
+          dispatch(successImageUpload(photo))
         }
         console.log(err)
       })
     }
   }
-  dispatch({ type: MEDIA_UPLOAD_END })
+  dispatch(endMediaUpload())
 }
